+++ Cql/Cql.Compiler/CqlOperatorsBinding.Bind.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using Hl7.Cql.Operators;
+using Hl7.Cql.Primitives;
 
 namespace Hl7.Cql.Compiler;
 
@@ -26,20 +26,21 @@ partial class CqlOperatorsBinding
         Expression[] args = arguments; // So we don't modify the original array
         var methods = ICqlOperators_MethodInfos_By_Name[methodName];
 
-        bool removedLastNullArgOnce = false;
-        while (true)
+        for (int i = 0; i < 2; i++) // Try twice, first with all arguments, then without the last one
         {
             Expression[] bindArgs = new Expression[args.Length];
 
             foreach (var method in methods)
             {
-                for (int argIndexForGenericMethod = 0; argIndexForGenericMethod < Math.Min(args.Length, 2); argIndexForGenericMethod++)
+                for (int argIndexForGenericMethod = 0;
+                     argIndexForGenericMethod < Math.Min(args.Length, 2);
+                     argIndexForGenericMethod++) // Try to get generic type from argument up to the second one
                 {
                     MethodInfo bindMethod;
                     if (method is { IsGenericMethodDefinition: true })
                     {
                         if (!args[argIndexForGenericMethod].Type.IsGenericType)
-                            continue;
+                            continue; // Not a generic argument, try again
 
                         var genericType = args[argIndexForGenericMethod].Type.GetGenericArguments()[0];
                         bindMethod = method.MakeGenericMethod(genericType);
@@ -71,11 +72,10 @@ partial class CqlOperatorsBinding
                 }
             }
 
-            if (!removedLastNullArgOnce && args[^1] is ConstantExpression { Value: null })
+            if (i == 0 && args[^1] is ConstantExpression { Value: null })
             {
                 // Handles precision cases where the last argument might be supplied or not
                 args = args[..^1];
-                removedLastNullArgOnce = true;
                 continue;
             }
 
@@ -114,4 +114,8 @@ partial class CqlOperatorsBinding
         MethodInfo method,
         params Expression[] expressions) =>
         Expression.Call(CqlContextExpressions.Operators_PropertyExpression, method, expressions);
+
+    private static bool IsInterval(Type type) =>
+        type.IsGenericType
+        && type.GetGenericTypeDefinition() == typeof(CqlInterval<>);
 }
